Move functions into TaxableTrade
