+++ src/calculator/taxable_trade.rs
@@ -1,9 +1,15 @@
 use crate::calculator::Currency;
+use crate::calculator::cost_book::CostBook;
 use crate::calculator::money::Money;
+use crate::calculator::trade::{Direction, Trade};
+use crate::{Config, skatteverket, writer};
+use anyhow::{anyhow, Result};
+use log::debug;
 use rust_decimal::Decimal;
 use rust_decimal_macros::dec;
 use serde::ser::SerializeStruct;
 use serde::{Serialize, Serializer};
+use std::collections::HashSet;
 
 // 1. Bought Crypto 1 from SEK      (cost in SEK),  sold to SEK      (sales in SEK)
 // 2. Bought Crypto 1 from SEK      (cost in SEK),  sold to Crypto 2 (SEK price as sales)
@@ -43,7 +49,7 @@ impl TaxableTrade {
         income: Money,
         costs: Vec<Money>,
         net_income: Option<Decimal>,
-    ) -> TaxableTrade {
+    ) -> Self {
         TaxableTrade {
             date,
             currency,
@@ -63,6 +69,73 @@ impl TaxableTrade {
             self.costs.iter()
                 .fold("".to_string(), |acc, c| format!("{}, {}", acc, c))
         }
+    }
+
+    pub(crate) async fn all_taxable_trades(trades: &Vec<Trade>) -> Vec<TaxableTrade> {
+        let mut unique_pairs: HashSet<(Currency, Currency)> = HashSet::new();
+
+        for t in trades {
+            let pair = (t.paid_currency.clone(), t.exchanged_currency.clone());
+            unique_pairs.insert(pair);
+        }
+
+        let mut taxable_trades: Vec<TaxableTrade> = vec![];
+        for pair in unique_pairs {
+            let result = Self::taxable_trades(&trades, &pair.0, &pair.1).await.unwrap();
+            taxable_trades.extend(result);
+        }
+
+        taxable_trades
+    }
+
+    pub(crate) async fn taxable_trades(trades: &Vec<Trade>, currency: &Currency, base_currency: &Currency) -> Result<Vec<TaxableTrade>> {
+        let book = CostBook::new(currency.clone(), base_currency.clone());
+
+        let (taxable_trades, book) =
+            trades.iter()
+                .fold((vec![], book), |(mut acc, mut book), trade| {
+                    let currency_match =
+                        trade.paid_currency.eq(&book.currency)
+                            && trade.exchanged_currency.eq(&book.base_currency);
+                    if currency_match {
+                        match trade.direction {
+                            Direction::Buy =>
+                                book.add_buy(trade),
+                            Direction::Sell => {
+                                let taxable_trade = book.add_sell(trade).unwrap();
+                                acc.push(taxable_trade);
+                            }
+                        }
+                    }
+                    (acc, book)
+                });
+
+        debug!("Remaining costs for {:?}:", book.currency);
+        book.costs.iter().for_each(|c| debug!("{:?}", c));
+        debug!("Taxable transactions:");
+        taxable_trades.iter().for_each(|t| debug!("{:?}", t));
+
+        Ok(taxable_trades)
+    }
+
+    pub(crate) async fn print_taxable_trades(taxable_trades: &Vec<TaxableTrade>, config: &Config) -> Result<()> {
+        if let Some(sru_conf) = &config.sru_file_config {
+            Self::print_sru_file(&taxable_trades, sru_conf.sru_org_num.clone(), sru_conf.sru_org_name.clone()).await?;
+        } else {
+            writer::print_csv_rows(&taxable_trades).await?;
+        }
+        Ok(())
+    }
+
+    async fn print_sru_file(taxable_trades: &Vec<TaxableTrade>, org_num: String, name: Option<String>) -> Result<()> {
+        let sru_file = skatteverket::SruFile::try_new(taxable_trades, org_num, name)
+            .ok_or(anyhow!("Failed to create SRU file from taxable trades"))?;
+
+        let stdout = std::io::stdout();
+        let handle = stdout.lock();
+        sru_file.write(handle)?;
+
+        Ok(())
     }
 }
